+++ src/main/java/cn/fql/fishbone/dao/mapper/RoleMapper.java
@@ -15,7 +15,7 @@ public interface RoleMapper {
 
     void updateRole(Role role);
 
-    void deleteRole(Integer id);
+    void deleteRole(Long id);
 
     List<Role> queryRole(RoleParam RoleParam);
 }
